requirements.txt
torch
transformers
datasets
sentence-transformers
fastapi
uvicorn[standard]
streamlit
pydantic
requests
python-dotenv
newspaper3k
beautifulsoup4
psycopg2-binary 
elasticsearch
captum
scikit-learn
pandas
numpy
gunicorn
feedparser
